enhance(load-gen): move function-specific config, setup and executor to seperate files to increase modularity - improve documentation - rearrange a bunch of stuff
